+++ src/bes/lims/exceptions.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+class TooManyRecordsError(Exception):
+    """Exception for when too many records where requested
+    """